requirements.txt
python-telegram-bot==20.7
pillow==10.1.0
pymupdf==1.23.8